refactor(bridge): extract StreamCodec
